Refactor(utils): Complete tokenize move (source, types, tests, docs)
Carries the function body, type, tests, benches, API-doc removal, and skill
update that pair with the example rename in 67cb0d5. The prior commit's git add
aborted on the already-moved example path and staged only the rename.
