+++ docs/src/content/examples/utils-tokenize.mdx
@@ -3,9 +3,9 @@ title: 'tokenize'
 id: 'utils-tokenize'
 exampleType: 'log'
 category: 'Utils'
-subcategory: 'Crypto'
+subcategory: 'Strings'
 description: 'Convert strings to lowercase tokens with hyphens'
-tags: ['utils', 'crypto', 'tokenize', 'normalize']
+tags: ['utils', 'strings', 'tokenize', 'normalize', 'slug']
 selectedFile: 'index.js'
 tip: 'Converts spaces and underscores to hyphens, removes special characters'
 ---
